cleanup of azure-related code
