+++ dep_tools/azure.py
@@ -1,11 +1,17 @@
+from functools import partial
+import io
+from multiprocessing.dummy import Pool as ThreadPool
 import os
 from pathlib import Path
 from typing import Union, Generator
 
 import azure.storage.blob
 from azure.storage.blob import ContainerClient
-from functools import partial
-from multiprocessing.dummy import Pool as ThreadPool
+import fiona
+from geopandas import GeoDataFrame
+from odc.geo.xr import to_cog
+from osgeo import gdal
+from xarray import DataArray, Dataset
 
 
 def get_container_client(
@@ -88,3 +94,51 @@ def build_vrt(
     vrt_file = f"data/{local_prefix}.vrt"
     gdal.BuildVRT(vrt_file, blobs, outputBounds=bounds)
     return Path(vrt_file)
+
+
+def write_to_blob_storage(
+    d: Union[DataArray, Dataset, GeoDataFrame, str],
+    path: Union[str, Path],
+    overwrite: bool = True,
+    use_odc_writer: bool = False,
+    client: ContainerClient | None = None,
+    **kwargs,
+) -> None:
+    # Allowing for a shared container client, which might be
+    # more efficient. If not provided, get one.
+    if client is None:
+        client = get_container_client()
+    blob_client = client.get_blob_client(str(path))
+    if not overwrite and blob_client.exists():
+        return
+
+    if isinstance(d, (DataArray, Dataset)):
+        if use_odc_writer:
+            if "driver" in kwargs:
+                del kwargs["driver"]
+            binary_data = to_cog(d, **kwargs)
+            blob_client.upload_blob(
+                binary_data, overwrite=overwrite, connection_timeout=TIMEOUT_SECONDS
+            )
+        else:
+            with io.BytesIO() as buffer:
+                # This is needed or rioxarray doesn't know what type it is writing
+                if "driver" not in kwargs:
+                    kwargs["driver"] = "COG"
+                d.rio.to_raster(buffer, **kwargs)
+                buffer.seek(0)
+                blob_client.upload_blob(
+                    buffer, overwrite=overwrite, connection_timeout=TIMEOUT_SECONDS
+                )
+
+    elif isinstance(d, GeoDataFrame):
+        with fiona.io.MemoryFile() as buffer:
+            d.to_file(buffer, **kwargs)
+            buffer.seek(0)
+            blob_client.upload_blob(buffer, overwrite=overwrite)
+    elif isinstance(d, str):
+        blob_client.upload_blob(d, overwrite=overwrite, **kwargs)
+    else:
+        raise ValueError(
+            "You can only write an Xarray DataArray or Dataset, or Geopandas GeoDataFrame"
+        )
